+++ src/reporter/health.py
@@ -2,7 +2,7 @@ from geocoding import geocoding
 from geocoding.factory import get_geo_cache, is_geo_coding_available
 from cache.factory import get_cache, is_cache_available
 from translators.factory import CRATE_BACKEND, TIMESCALE_BACKEND, \
-    default_backend
+    default_backend, Timescale_backend
 
 
 def check_db(db=CRATE_BACKEND):
@@ -16,6 +16,11 @@ def check_db(db=CRATE_BACKEND):
             health = trans.get_health()
             return health
     if db == TIMESCALE_BACKEND:
+        from translators.timescale import postgres_translator_instance
+        with postgres_translator_instance() as trans:
+            health = trans.get_health()
+            return health
+    if db2 == TIMESCALE_BACKEND:
         from translators.timescale import postgres_translator_instance
         with postgres_translator_instance() as trans:
             health = trans.get_health()
@@ -82,11 +87,14 @@ def get_health(with_geocoder=False):
 
     # Check defaultDB (critical)
     db = default_backend().lower()
+    db2 = Timescale_backend().lower()
     try:
         res = _check_critical(check_db(db), res, db)
+        res = _check_critical(check_db(db2), res, db2)
     except Exception:
         res['status'] = 'fail'
         res.setdefault('details', {})[db] = 'cannot reach ' + db
+        res.setdefault('details', {})[db2] = 'cannot reach ' + db2
 
     # TODO add not critical check if a secondary db is configured